move dispatch to internal directory
